// src/types/avatar.ts
export enum BatchAvatarStatus {
  NotStarted = "NotStarted",
  Running = "Running",
  Succeeded = "Succeeded",
  Failed = "Failed",
}

// Avaliable Region: https://learn.microsoft.com/en-us/azure/ai-services/speech-service/text-to-speech-avatar/what-is-text-to-speech-avatar#available-locations
export enum RealtimeAvatarSupportRegion {
  SoutheastAsia = "southeastasia",
  NorthEurope = "northeurope",
  WestEurope = "westeurope",
  SwedenCentral = "swedencentral",
  SouthCentralUS = "southcentralus",
  EastUS2 = "eastus2",
  WestUS2 = "westus2",
}